+++ src/components/CryptoList.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { NavLink } from 'react-router-dom';
-import rightArrow from './right-arrows.png';
+import rightArrow from './assets/right-arrows.png';
 
 function CryptoList(props) {
   const {
